pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["jd <judge.py@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
gunicorn = "^20.1.0"
python-epo-ops-client = "^4.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.19.0"
flake8 = "^4.0.1"
black = {version = "^22.6.0", allow-prereleases = true}
isort = "^5.10.1"
mypy = "^0.961"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79
